Updated to current working conditions

+++ database/migrations/2019_01_02_172319_change_data_to_json_amcharts.php
@@ -33,9 +33,15 @@ class ChangeDataToJsonAmcharts extends Migration
      */
     public function down()
     {
+        if (Schema::hasColumn('amcharts', 'series'))
+        {
+            Schema::table('amcharts', function (Blueprint $table) {
+                $table->dropColumn('series');
+            });
+        }
+
         Schema::table('amcharts', function (Blueprint $table) {
             $table->longText('data')->nullable()->change();
-            $table->dropColumn('series');
         });
     }
 }
